refactor: 숫자야구 유틸 함수 및 상수 분리
